+++ tests/fields/file_tests.py
@@ -14,6 +14,12 @@ from mongoengine import *
 from mongoengine.connection import get_db
 from mongoengine.python_support import PY3, b, StringIO
 
+try:
+    from PIL import Image
+    HAS_PIL = True
+except ImportError:
+    HAS_PIL = False
+
 TEST_IMAGE_PATH = os.path.join(os.path.dirname(__file__), 'mongoengine.png')
 TEST_IMAGE2_PATH = os.path.join(os.path.dirname(__file__), 'mongodb_leaf.png')
 
@@ -255,8 +261,8 @@ class FileTest(unittest.TestCase):
         self.assertFalse(test_file.the_file in [{"test": 1}])
 
     def test_image_field(self):
-        if PY3:
-            raise SkipTest('PIL does not have Python 3 support')
+        if not HAS_PIL:
+            raise SkipTest('PIL not installed')
 
         class TestImage(Document):
             image = ImageField()
@@ -278,8 +284,8 @@ class FileTest(unittest.TestCase):
         t.image.delete()
 
     def test_image_field_reassigning(self):
-        if PY3:
-            raise SkipTest('PIL does not have Python 3 support')
+        if not HAS_PIL:
+            raise SkipTest('PIL not installed')
 
         class TestFile(Document):
             the_file = ImageField()
@@ -294,8 +300,8 @@ class FileTest(unittest.TestCase):
         self.assertEqual(test_file.the_file.size, (45, 101))
 
     def test_image_field_resize(self):
-        if PY3:
-            raise SkipTest('PIL does not have Python 3 support')
+        if not HAS_PIL:
+            raise SkipTest('PIL not installed')
 
         class TestImage(Document):
             image = ImageField(size=(185, 37))
@@ -317,8 +323,8 @@ class FileTest(unittest.TestCase):
         t.image.delete()
 
     def test_image_field_resize_force(self):
-        if PY3:
-            raise SkipTest('PIL does not have Python 3 support')
+        if not HAS_PIL:
+            raise SkipTest('PIL not installed')
 
         class TestImage(Document):
             image = ImageField(size=(185, 37, True))
@@ -340,8 +346,8 @@ class FileTest(unittest.TestCase):
         t.image.delete()
 
     def test_image_field_thumbnail(self):
-        if PY3:
-            raise SkipTest('PIL does not have Python 3 support')
+        if not HAS_PIL:
+            raise SkipTest('PIL not installed')
 
         class TestImage(Document):
             image = ImageField(thumbnail_size=(92, 18))
@@ -388,6 +394,14 @@ class FileTest(unittest.TestCase):
         self.assertEqual(test_file.the_file.read(),
                           b('Hello, World!'))
 
+        test_file = TestFile.objects.first()
+        test_file.the_file = b('HELLO, WORLD!')
+        test_file.save()
+
+        test_file = TestFile.objects.first()
+        self.assertEqual(test_file.the_file.read(),
+                          b('HELLO, WORLD!'))
+
     def test_copyable(self):
         class PutFile(Document):
             the_file = FileField()
@@ -409,8 +423,8 @@ class FileTest(unittest.TestCase):
 
     def test_get_image_by_grid_id(self):
 
-        if PY3:
-            raise SkipTest('PIL does not have Python 3 support')
+        if not HAS_PIL:
+            raise SkipTest('PIL not installed')
 
         class TestImage(Document):