+++ tests/queryset/queryset.py
@@ -545,6 +545,15 @@ class QuerySetTest(unittest.TestCase):
         self.assertEqual("Bob", bob.name)
         self.assertEqual(30, bob.age)
 
+    def test_upsert_one(self):
+        self.Person.drop_collection()
+
+        self.Person.objects(name="Bob", age=30).update_one(upsert=True)
+
+        bob = self.Person.objects.first()
+        self.assertEqual("Bob", bob.name)
+        self.assertEqual(30, bob.age)
+
     def test_set_on_insert(self):
         self.Person.drop_collection()
 
@@ -3113,7 +3122,10 @@ class QuerySetTest(unittest.TestCase):
         self.assertEqual([], bars)
 
         self.assertRaises(ConfigurationError, Bar.objects,
-                            read_preference='Primary')
+                          read_preference='Primary')
+
+        bars = Bar.objects(read_preference=ReadPreference.SECONDARY_PREFERRED)
+        self.assertEqual(bars._read_preference, ReadPreference.SECONDARY_PREFERRED)
 
     def test_json_simple(self):